Removing scheduler, field_permissions, paragraphs_library, menu_ui, contextual_filter dependency and removes field_project_team from Project content type
Removes the dependency on scheduler module.
Removes the dependency on the field_permissions module.
Removes the dependency on the menu_ui module.
Removes the dependency on paragraphs_library module.
Also removes a number of declared dependencies from the .info.yml file that are not in fact needed.

Removes 'field_project_team' from the Project content type to eliminate a circular
dependency.